Added Dynamic file editing support
Made the whole Git repo be a "volume", which is a persistent file system
as opposed to copy it into the Container. Therefore, now the docker can
edit the Project, do git stuff and so on. Now I'm actually considering a
linux distro as a base because with that we could have ssh on it and
could commit and push from inside the distro. Other customization would
also be easier without this constrained python only environment.