Fix/multilabel confusion matrix (#51)
* Init: init fixing environment + scratch script

* Added binarized_target_decoder to dataloader

Function takes binarrized encoded target array and decodes it back.
useful for confusion matrix.

* Confusion Matrix working for all except metals

* Fixed problem with metals flag

The one dimensional stuff wasn't working due to some places expecting
lists of lists and only getting lists. Fixed now

* Chore: get rid of ported testing scripts

* Chore: restore model_evaluation as much as possible

* Chore: get rid of TODO comments fixed by dataloader

* Fix: added the dataloader y_labels to the evaluation + cleanup

Dataloader was added to evaluation to label the confusion matrix. If
this is too much mixing of the dataloader we can also of course just
pass the decoded y_pred and y_test.

* Refactor train_script and multi dim support for evaluation and plotting

Some setup for adding the multi dimensional support for the Confusion
matrix the way it produces multiple cms for each target of the
--targets.

* Feature: get_target_columns_separated() added to dataloader

This function returns a list of list of the columns of each target. This
is needed to make multiple confusion matrices.

* Add support for multiple cm types

Added support for one target, one dim target array and
multiple targets, multidimensional target array
single target, multidimensional target array for confusion matrices.

* Hotfix: remove to_csv("yeet")

---------

Co-authored-by: TiagoW <twuerthner@ethz.ch>